requirements.txt
edge-tts
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
PyYAML==6.0.3
google-genai==1.56.0
